Extended the playfield 32 pixels beyond the screen.
I initially tried to mirror the sprites (as the origional asteroids), which
I hacked together, but the bounding box collisions only checked the bottom/right
half of the sprite when its wrapped.

I *could* mirror the entities x/y position and then check bounding box
when the entity is wrapped but I'm already pressed for CPU time
that it would take slowdown to 35 entities, which is too little for me.